+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## CHANGELOG
+### 4.x.x (In progress)
+**SDK**
+- **Feature:** Added interface `MiniAppFileDownloader` to download files in device from miniapp.
 
 ### 4.0.0 (2022-03-01)
-
 **SDK**
 - **Update:** Updated create method with optional fromCache variable that helps to load the mini-app from cache.
 - **Upgraded:** Target SDK is now Android 12 (API level 31)